Updating to head

+++ includes/api/legacy/v3/class-wc-api-products.php
@@ -177,7 +177,7 @@ class WC_API_Products extends WC_API_Resource {
 	 * @since 2.1
 	 * @param int $id the product ID
 	 * @param string $fields
-	 * @return array
+	 * @return array|WP_Error
 	 */
 	public function get_product( $id, $fields = null ) {
 
@@ -222,9 +222,11 @@ class WC_API_Products extends WC_API_Resource {
 	 * Get the total number of products
 	 *
 	 * @since 2.1
+	 *
 	 * @param string $type
 	 * @param array $filter
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function get_products_count( $type = null, $filter = array() ) {
 		try {
@@ -248,8 +250,10 @@ class WC_API_Products extends WC_API_Resource {
 	 * Create a new product.
 	 *
 	 * @since 2.2
+	 *
 	 * @param array $data posted data
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function create_product( $data ) {
 		$id = 0;
@@ -292,13 +296,13 @@ class WC_API_Products extends WC_API_Resource {
 			$post_content = isset( $data['description'] ) ? wc_clean( $data['description'] ) : '';
 			if ( $post_content && isset( $data['enable_html_description'] ) && true === $data['enable_html_description'] ) {
 
-				$post_content = $data['description'];
+				$post_content = wp_filter_post_kses( $data['description'] );
 			}
 
 			// Enable short description html tags.
 			$post_excerpt = isset( $data['short_description'] ) ? wc_clean( $data['short_description'] ) : '';
 			if ( $post_excerpt && isset( $data['enable_html_short_description'] ) && true === $data['enable_html_short_description'] ) {
-				$post_excerpt = $data['short_description'];
+				$post_excerpt = wp_filter_post_kses( $data['short_description'] );
 			}
 
 			$classname = WC_Product_Factory::get_classname_from_product_type( $data['type'] );
@@ -361,9 +365,11 @@ class WC_API_Products extends WC_API_Resource {
 	 * Edit a product
 	 *
 	 * @since 2.2
+	 *
 	 * @param int $id the product ID
 	 * @param array $data
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function edit_product( $id, $data ) {
 		try {
@@ -401,14 +407,14 @@ class WC_API_Products extends WC_API_Resource {
 			// Product short description.
 			if ( isset( $data['short_description'] ) ) {
 				// Enable short description html tags.
-				$post_excerpt = ( isset( $data['enable_html_short_description'] ) && true === $data['enable_html_short_description'] ) ? $data['short_description'] : wc_clean( $data['short_description'] );
+				$post_excerpt = ( isset( $data['enable_html_short_description'] ) && true === $data['enable_html_short_description'] ) ? wp_filter_post_kses( $data['short_description'] ) : wc_clean( $data['short_description'] );
 				$product->set_short_description( $post_excerpt );
 			}
 
 			// Product description.
 			if ( isset( $data['description'] ) ) {
 				// Enable description html tags.
-				$post_content = ( isset( $data['enable_html_description'] ) && true === $data['enable_html_description'] ) ? $data['description'] : wc_clean( $data['description'] );
+				$post_content = ( isset( $data['enable_html_description'] ) && true === $data['enable_html_description'] ) ? wp_filter_post_kses( $data['description'] ) : wc_clean( $data['description'] );
 				$product->set_description( $post_content );
 			}
 
@@ -459,9 +465,11 @@ class WC_API_Products extends WC_API_Resource {
 	 * Delete a product.
 	 *
 	 * @since 2.2
+	 *
 	 * @param int $id the product ID.
 	 * @param bool $force true to permanently delete order, false to move to trash.
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function delete_product( $id, $force = false ) {
 
@@ -480,18 +488,23 @@ class WC_API_Products extends WC_API_Resource {
 			if ( $product->is_type( 'variable' ) ) {
 				foreach ( $product->get_children() as $child_id ) {
 					$child = wc_get_product( $child_id );
-					$child->delete( true );
+					if ( ! empty( $child ) ) {
+						$child->delete( true );
+					}
 				}
-			} elseif ( $product->is_type( 'grouped' ) ) {
+			} else {
+				// For other product types, if the product has children, remove the relationship.
 				foreach ( $product->get_children() as $child_id ) {
 					$child = wc_get_product( $child_id );
-					$child->set_parent_id( 0 );
-					$child->save();
+					if ( ! empty( $child ) ) {
+						$child->set_parent_id( 0 );
+						$child->save();
+					}
 				}
 			}
 
 			$product->delete( true );
-			$result = $product->get_id() > 0 ? false : true;
+			$result = ! ( $product->get_id() > 0 );
 		} else {
 			$product->delete();
 			$result = 'trash' === $product->get_status();
@@ -521,7 +534,7 @@ class WC_API_Products extends WC_API_Resource {
 	 * @since 2.1
 	 * @param int $id the product ID to get reviews for
 	 * @param string $fields fields to include in response
-	 * @return array
+	 * @return array|WP_Error
 	 */
 	public function get_product_reviews( $id, $fields = null ) {
 
@@ -556,10 +569,10 @@ class WC_API_Products extends WC_API_Resource {
 	 * @since 2.4.0
 	 * @param int $id the product ID to get orders for
 	 * @param string fields  fields to retrieve
-	 * @param string $filter filters to include in response
+	 * @param array $filter filters to include in response
 	 * @param string $status the order status to retrieve
 	 * @param $page  $page   page to retrieve
-	 * @return array
+	 * @return array|WP_Error
 	 */
 	public function get_product_orders( $id, $fields = null, $filter = array(), $status = null, $page = 1 ) {
 		global $wpdb;
@@ -594,8 +607,10 @@ class WC_API_Products extends WC_API_Resource {
 	 * Get a listing of product categories
 	 *
 	 * @since 2.2
+	 *
 	 * @param string|null $fields fields to limit response to
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function get_product_categories( $fields = null ) {
 		try {
@@ -622,9 +637,11 @@ class WC_API_Products extends WC_API_Resource {
 	 * Get the product category for the given ID
 	 *
 	 * @since 2.2
+	 *
 	 * @param string $id product category term ID
 	 * @param string|null $fields fields to limit response to
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function get_product_category( $id, $fields = null ) {
 		try {
@@ -861,8 +878,10 @@ class WC_API_Products extends WC_API_Resource {
 	 * Get a listing of product tags.
 	 *
 	 * @since  2.5.0
+	 *
 	 * @param  string|null $fields Fields to limit response to
-	 * @return array               Product tags
+	 *
+	 * @return array|WP_Error
 	 */
 	public function get_product_tags( $fields = null ) {
 		try {
@@ -889,9 +908,11 @@ class WC_API_Products extends WC_API_Resource {
 	 * Get the product tag for the given ID.
 	 *
 	 * @since  2.5.0
+	 *
 	 * @param  string $id          Product tag term ID
 	 * @param  string|null $fields Fields to limit response to
-	 * @return array               Product tag
+	 *
+	 * @return array|WP_Error
 	 */
 	public function get_product_tag( $id, $fields = null ) {
 		try {
@@ -1123,11 +1144,16 @@ class WC_API_Products extends WC_API_Resource {
 	 *
 	 * @since 2.1
 	 * @param WC_Product|int $product
-	 * @return WC_Product
+	 *
+	 * @return array
 	 */
 	private function get_product_data( $product ) {
 		if ( is_numeric( $product ) ) {
 			$product = wc_get_product( $product );
+		}
+
+		if ( ! is_a( $product, 'WC_Product' ) ) {
+			return array();
 		}
 
 		return array(
@@ -1515,22 +1541,17 @@ class WC_API_Products extends WC_API_Resource {
 
 		} else {
 
-			// Regular Price
+			// Regular Price.
 			if ( isset( $data['regular_price'] ) ) {
 				$regular_price = ( '' === $data['regular_price'] ) ? '' : $data['regular_price'];
-			} else {
-				$regular_price = $product->get_regular_price();
+				$product->set_regular_price( $regular_price );
 			}
 
-			// Sale Price
+			// Sale Price.
 			if ( isset( $data['sale_price'] ) ) {
 				$sale_price = ( '' === $data['sale_price'] ) ? '' : $data['sale_price'];
-			} else {
-				$sale_price = $product->get_sale_price();
+				$product->set_sale_price( $sale_price );
 			}
-
-			$product->set_regular_price( $regular_price );
-			$product->set_sale_price( $sale_price );
 
 			if ( isset( $data['sale_price_dates_from'] ) ) {
 				$date_from = $data['sale_price_dates_from'];
@@ -1550,10 +1571,11 @@ class WC_API_Products extends WC_API_Resource {
 
 			$product->set_date_on_sale_to( $date_to );
 			$product->set_date_on_sale_from( $date_from );
-			if ( $product->is_on_sale() ) {
-				$product->set_price( $product->get_sale_price() );
+
+			if ( $product->is_on_sale( 'edit' ) ) {
+				$product->set_price( $product->get_sale_price( 'edit' ) );
 			} else {
-				$product->set_price( $product->get_regular_price() );
+				$product->set_price( $product->get_regular_price( 'edit' ) );
 			}
 		}
 
@@ -1742,9 +1764,11 @@ class WC_API_Products extends WC_API_Resource {
 	 * Save variations.
 	 *
 	 * @since  2.2
+	 *
 	 * @param  WC_Product $product
 	 * @param  array $request
-	 * @return WC_Product
+	 *
+	 * @return bool
 	 * @throws WC_API_Exception
 	 */
 	protected function save_variations( $product, $request ) {
@@ -1979,11 +2003,9 @@ class WC_API_Products extends WC_API_Resource {
 
 		// Shipping class
 		if ( isset( $data['shipping_class'] ) ) {
-			$data_store         = $product->get_data_store();
-			$shipping_class_id  = $data_store->get_shipping_class_id_by_slug( wc_clean( $data['shipping_class'] ) );
-			if ( $shipping_class_id ) {
-				$product->set_shipping_class_id( $shipping_class_id );
-			}
+			$data_store        = $product->get_data_store();
+			$shipping_class_id = $data_store->get_shipping_class_id_by_slug( wc_clean( $data['shipping_class'] ) );
+			$product->set_shipping_class_id( $shipping_class_id );
 		}
 
 		return $product;
@@ -2014,7 +2036,7 @@ class WC_API_Products extends WC_API_Resource {
 			}
 
 			$download = new WC_Product_Download();
-			$download->set_id( $key );
+			$download->set_id( ! empty( $file['id'] ) ? $file['id'] : wp_generate_uuid4() );
 			$download->set_name( $file['name'] ? $file['name'] : wc_get_filename_from_url( $file['file'] ) );
 			$download->set_file( apply_filters( 'woocommerce_file_download_path', $file['file'], $product, $key ) );
 			$files[]  = $download;
@@ -2204,7 +2226,7 @@ class WC_API_Products extends WC_API_Resource {
 	 * @since 2.5.0
 	 * @param string $image_url
 	 * @param string $upload_for
-	 * @return int|WP_Error Attachment id
+	 * @return array|WP_Error
 	 */
 	protected function upload_image_from_url( $image_url, $upload_for = 'product_image' ) {
 		$file_name = basename( current( explode( '?', $image_url ) ) );
@@ -2422,8 +2444,10 @@ class WC_API_Products extends WC_API_Resource {
 	 * Get a listing of product attributes
 	 *
 	 * @since 2.5.0
+	 *
 	 * @param string|null $fields fields to limit response to
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function get_product_attributes( $fields = null ) {
 		try {
@@ -2456,9 +2480,11 @@ class WC_API_Products extends WC_API_Resource {
 	 * Get the product attribute for the given ID
 	 *
 	 * @since 2.5.0
+	 *
 	 * @param string $id product attribute term ID
 	 * @param string|null $fields fields to limit response to
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function get_product_attribute( $id, $fields = null ) {
 		global $wpdb;
@@ -2543,8 +2569,10 @@ class WC_API_Products extends WC_API_Resource {
 	 * Create a new product attribute.
 	 *
 	 * @since 2.5.0
+	 *
 	 * @param array $data Posted data.
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function create_product_attribute( $data ) {
 		global $wpdb;
@@ -2624,9 +2652,11 @@ class WC_API_Products extends WC_API_Resource {
 	 * Edit a product attribute.
 	 *
 	 * @since 2.5.0
+	 *
 	 * @param int $id the attribute ID.
 	 * @param array $data
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function edit_product_attribute( $id, $data ) {
 		global $wpdb;
@@ -2706,8 +2736,10 @@ class WC_API_Products extends WC_API_Resource {
 	 * Delete a product attribute.
 	 *
 	 * @since  2.5.0
+	 *
 	 * @param  int $id the product attribute ID.
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function delete_product_attribute( $id ) {
 		global $wpdb;
@@ -2766,9 +2798,11 @@ class WC_API_Products extends WC_API_Resource {
 	 * Get a listing of product attribute terms.
 	 *
 	 * @since 2.5.0
+	 *
 	 * @param int $attribute_id Attribute ID.
 	 * @param string|null $fields Fields to limit response to.
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function get_product_attribute_terms( $attribute_id, $fields = null ) {
 		try {
@@ -2823,10 +2857,12 @@ class WC_API_Products extends WC_API_Resource {
 	 * Get the product attribute term for the given ID.
 	 *
 	 * @since 2.5.0
+	 *
 	 * @param int $attribute_id Attribute ID.
 	 * @param string $id Product attribute term ID.
 	 * @param string|null $fields Fields to limit response to.
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function get_product_attribute_term( $attribute_id, $id, $fields = null ) {
 		global $wpdb;
@@ -2873,9 +2909,11 @@ class WC_API_Products extends WC_API_Resource {
 	 * Create a new product attribute term.
 	 *
 	 * @since 2.5.0
+	 *
 	 * @param int $attribute_id Attribute ID.
 	 * @param array $data Posted data.
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function create_product_attribute_term( $attribute_id, $data ) {
 		global $wpdb;
@@ -2935,10 +2973,12 @@ class WC_API_Products extends WC_API_Resource {
 	 * Edit a product attribute term.
 	 *
 	 * @since 2.5.0
+	 *
 	 * @param int $attribute_id Attribute ID.
 	 * @param int $id the attribute ID.
 	 * @param array $data
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function edit_product_attribute_term( $attribute_id, $id, $data ) {
 		global $wpdb;
@@ -2994,9 +3034,11 @@ class WC_API_Products extends WC_API_Resource {
 	 * Delete a product attribute term.
 	 *
 	 * @since  2.5.0
+	 *
 	 * @param int $attribute_id Attribute ID.
 	 * @param int $id the product attribute ID.
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function delete_product_attribute_term( $attribute_id, $id ) {
 		global $wpdb;
@@ -3032,6 +3074,8 @@ class WC_API_Products extends WC_API_Resource {
 
 	/**
 	 * Clear product
+	 *
+	 * @param int $product_id
 	 */
 	protected function clear_product( $product_id ) {
 		if ( ! is_numeric( $product_id ) || 0 >= $product_id ) {
@@ -3060,8 +3104,10 @@ class WC_API_Products extends WC_API_Resource {
 	 * WC_API_Products->create_product() and WC_API_Products->edit_product()
 	 *
 	 * @since 2.4.0
+	 *
 	 * @param array $data
-	 * @return array
+	 *
+	 * @return array|WP_Error
 	 */
 	public function bulk( $data ) {
